+++ src/pages/auth/register/index.tsx
@@ -1,4 +1,4 @@
-import SignupForm from '@/features/auth/signup/signupForm';
+import SignupForm from '@/features/auth/register/registerForm';
 import useSignupForm from '@/shared/hooks/useSignupForm';
 import Header from '@/widgets/Header';
 
@@ -11,7 +11,7 @@ export function Register() {
       <Header title="회원목록" />
       <div className="mt-[22.5px] flex w-full justify-center md:mt-[80px]">
         <SignupForm
-          signupFormValue={signupFormValue}
+          registerFormValue={signupFormValue}
           handleSubmit={handleSubmit}
           handleChange={handleChange}
           errors={errors}
